refactor: change price field from Integer to BigDecimal across Food and DTOs
Updated the price field in Food entity, FoodRequestDTO, and FoodResponseDTO
from Integer to BigDecimal to support decimal precision in monetary values.

+++ src/main/java/vendas/Vendedor/Food/FoodRequestDTO.java
@@ -1,4 +1,6 @@
 package vendas.Vendedor.Food;
 
-public record FoodRequestDTO(String title, String image, Integer price) {
+import java.math.BigDecimal;
+
+public record FoodRequestDTO(String title, String image, BigDecimal price) {
 }
